refactor(api-headless-cms): extract references testing and setting up the models
